@@ -3,18 +3,19 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.FileProviders;
+using Microsoft.Extensions.Hosting;
 using Microsoft.SourceBrowser.SourceIndexServer.Models;
 
 namespace Microsoft.SourceBrowser.SourceIndexServer
 {
     public class Startup
     {
-        public Startup(IHostingEnvironment env)
+        public Startup(IWebHostEnvironment env)
         {
             Environment = env;
         }
 
-        public IHostingEnvironment Environment { get; }
+        public IWebHostEnvironment Environment { get; }
 
         // This method gets called by the runtime. Use this method to add services to the container.
         // For more information on how to configure your application, visit https://go.microsoft.com/fwlink/?LinkID=398940
@@ -29,13 +30,14 @@ namespace Microsoft.SourceBrowser.SourceIndexServer
             }
 
             services.AddSingleton(new Index(RootPath));
-            services.AddMvc();
+            services.AddControllersWithViews();
+            services.AddRazorPages();
         }
 
         public string RootPath { get; set; }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             app.Use(async (context, next) =>
             {
@@ -54,7 +56,14 @@ namespace Microsoft.SourceBrowser.SourceIndexServer
                 FileProvider = new PhysicalFileProvider(RootPath),
             });
             app.UseStaticFiles();
-            app.UseMvc();
+            app.UseRouting();
+
+
+            app.UseEndpoints(endPoints =>
+            {
+                endPoints.MapRazorPages();
+                endPoints.MapControllers();
+            });
         }
     }
 }
